Added libev

+++ build-package.sh
@@ -6,6 +6,12 @@ usage() {
   exit 1
 }
 
+QUICK=no
+if [ "$1" == "-q" ]; then
+  shift
+  QUICK=yes
+fi
+
 PACKAGE=$1
 [ ! -z "${PACKAGE}" ] || usage
 
@@ -14,10 +20,16 @@ cd "$(dirname $0)"
 cd "$(pwd -P)/${PACKAGE}"
 
 docker build -t "insideo/uc-linux-${PACKAGE}-build" --pull .
-docker rm -f "uc-linux-${PACKAGE}-tmp" 2>/dev/null || true
-docker run --name "uc-linux-${PACKAGE}-tmp" "insideo/uc-linux-${PACKAGE}-build" /bin/sh
-docker cp "uc-linux-${PACKAGE}-tmp:/packages" ..
-docker rm -f "uc-linux-${PACKAGE}-tmp"
-docker rmi "insideo/uc-linux-${PACKAGE}-build"
+
+if [ "${QUICK}" != "yes" ]; then
+  docker rm -f "uc-linux-${PACKAGE}-tmp" 2>/dev/null || true
+  docker run --name "uc-linux-${PACKAGE}-tmp" "insideo/uc-linux-${PACKAGE}-build" /bin/sh
+  docker cp "uc-linux-${PACKAGE}-tmp:/packages" ..
+  docker rm -f "uc-linux-${PACKAGE}-tmp"
+  docker rmi "insideo/uc-linux-${PACKAGE}-build"
+else 
+  echo "Quick build of package ${PACKAGE} complete."
+  exit 0
+fi
 
 echo "Build of package ${PACKAGE} complete."
